[PWA] ngsw config
- setting maxAge:3h with strategy:performance, will remove the file 
after 3h, so it will not be available offline after 3h.
so we changed it to a longer maxAge with strategy:freshness.
this will make the file available offline, but every time these urls 
will be requested from the network.
files excecluded (!/**/*.*) i.e has extension

- updateMode for assets changed to lazy, so we don't need to update the 
cache until it requested